refactor: move RequestTimeoutError and StaleRequestError to util/errors.ts
Error classes defined from Error constructor belong in util/errors.ts
alongside the other error types. Connection.ts now re-exports them
for backward compatibility; BaseSession and tests import directly from
util/errors.
